requirements.txt
flask
flask-sqlalchemy
flask-login
requests
fpdf2
pypdf
python-dotenv
werkzeug